requirements.txt
django-stringfield
sorl-thumbnail
django-extensions
